+++ LibCarla/source/carla/multigpu/primary.h
@@ -14,7 +14,7 @@
 #include "carla/streaming/detail/Message.h"
 #include "carla/multigpu/listener.h"
 
-#include <boost/asio/deadline_timer.hpp>
+#include <boost/asio/steady_timer.hpp>
 #include <boost/asio/io_context.hpp>
 #include <boost/asio/ip/tcp.hpp>
 #include <boost/asio/strand.hpp>
@@ -92,7 +92,7 @@ namespace multigpu {
 
     time_duration _timeout;
 
-    boost::asio::deadline_timer _deadline;
+    boost::asio::steady_timer _deadline;
 
     boost::asio::io_context::strand _strand;