+++ LibCarla/source/carla/streaming/EndPoint.h
@@ -75,13 +75,13 @@ namespace detail {
   static inline auto make_address(const std::string &address) {
     boost::asio::io_context io_context;
     boost::asio::ip::tcp::resolver resolver(io_context);
-    boost::asio::ip::tcp::resolver::query query(boost::asio::ip::tcp::v4(), address, "", boost::asio::ip::tcp::resolver::query::canonical_name);
-    boost::asio::ip::tcp::resolver::iterator iter = resolver.resolve(query);
-    boost::asio::ip::tcp::resolver::iterator end;
-    while (iter != end)
-    {
-      boost::asio::ip::tcp::endpoint endpoint = *iter++;
-      return endpoint.address();
+    auto results = resolver.resolve(
+        boost::asio::ip::tcp::v4(),
+        address,
+        std::string(),
+        boost::asio::ip::resolver_base::canonical_name);
+    for (const auto &result : results) {
+      return result.endpoint().address();
     }
     return boost::asio::ip::make_address(address);
   }