+++ LibCarla/source/carla/streaming/detail/tcp/Client.cpp
@@ -13,6 +13,8 @@
 #include "carla/Time.h"
 
 #include <boost/asio/connect.hpp>
+
+#include <chrono>
 #include <boost/asio/read.hpp>
 #include <boost/asio/write.hpp>
 #include <boost/asio/bind_executor.hpp>
@@ -150,7 +152,7 @@ namespace tcp {
 
   void Client::Reconnect() {
     auto self = shared_from_this();
-    _connection_timer.expires_from_now(time_duration::seconds(1u));
+    _connection_timer.expires_after(std::chrono::seconds(1));
     _connection_timer.async_wait([this, self](boost::system::error_code ec) {
       if (!ec) {
         Connect();