+++ LibCarla/source/test/common/test_streaming.cpp
@@ -27,17 +27,18 @@ public:
   boost::asio::io_context service;
 
   explicit io_context_running(size_t threads = 2u)
-    : _work_to_do(service) {
+    : _work_to_do(boost::asio::make_work_guard(service)) {
     _threads.CreateThreads(threads, [this]() { service.run(); });
   }
 
   ~io_context_running() {
+    _work_to_do.reset();
     service.stop();
   }
 
 private:
 
-  boost::asio::io_context::work _work_to_do;
+  boost::asio::executor_work_guard<boost::asio::io_context::executor_type> _work_to_do;
 
   carla::ThreadGroup _threads;
 };