+++ Util/BuildTools/BuildOSMRenderer.sh
@@ -68,8 +68,8 @@ echo "Building osm-map-renderer"
 mkdir -p ${OSM_RENDERER_BUILD}
 cd ${OSM_RENDERER_BUILD}
 
-cmake -DCMAKE_CXX_FLAGS="-std=c++17 -g -pthread -I${CARLA_BUILD_FOLDER}/boost-1.84.0-c10-install/include" \
--DCMAKE_EXE_LINKER_FLAGS="-L${CARLA_BUILD_FOLDER}/boost-1.84.0-c10-install/lib -lboost_system" \
+cmake -DCMAKE_CXX_FLAGS="-std=c++17 -g -pthread -I${CARLA_BUILD_FOLDER}/boost-1.90.0-c10-install/include" \
+-DCMAKE_EXE_LINKER_FLAGS="-L${CARLA_BUILD_FOLDER}/boost-1.90.0-c10-install/lib -lboost_system" \
     ${OSM_RENDERER_SOURCE}
 make -j$(nproc)